Fix to DS9FX mod conflicting with phased techs
As you may be aware, the current HandleNoDamageThroughShields makes it so absolutely no weapon can bypass the shields. This is fine at first. Problem comes for when a weapon is actually meant to do that like the Phased Plasma Torpedoes (including the stock ones). Those techs work on the concept of the torpedo "bypassing" the shield and dealing the stat and visual damage they would apply normally, so the torpedo actually hits the hull and deals the damage to the ship as if the shield was not there. The handler as it was previously made all those mods (as well as the stock function) inoperable. Before I knew DS9FX was being updated to newer versions, I made a patch for the 1.0 DS9FX file, but knowing how this project is still very much alive (thank you for informing me), I decided to just do the update here as a commit to fix the issue permanently and without breaking someone's different DS9FX version.

IMPORTANT NOTE: The phased weaponry is NOT how drones from the stargate pack work, those use a far inferior mock which only targets a single subsystem at a time and never deals visible damage. The real phased weaponry works as how the literal torpedo would have acted if there was no shield on the first place.